+++ src/modules/bgckmicrowave_mod.f90
@@ -2874,24 +2874,24 @@ contains
   end subroutine bennartz
 
   !--------------------------------------------------------------------------
-  ! atmsTest1Flagbit7Check
+  ! atmsMwhs2Test1Flagbit7Check
   !--------------------------------------------------------------------------
-  subroutine atmsTest1Flagbit7Check (itest, KCANO, KMARQ, KNOSAT, ICHECK, KNO, KNT, STNID,  B7CHCK)
+  subroutine atmsMwhs2Test1Flagbit7Check (itest, KCANO, KMARQ, KNOSAT, ICHECK, KNO, KNT, STNID,  B7CHCK)
 
-    !:Purpose:               1) test 1: Check flag bit 7 on from the first bgckAtms program
+    !:Purpose:               1) test 1: Check flag bit 7 on from the first bgckAtms/bgckMwhs2 program
     !                        Includes observations flagged for cloud liquid water, scattering index,
     !                        dryness index plus failure of several QC checks.
 
 
     ! Arguments
     integer,     intent(in)                                   :: itest(:)                 ! test number
-    integer,     intent(in)                                   :: KCANO(KNO,KNT)                 ! observations channels
+    integer,     intent(in)                                   :: KCANO(KNO,KNT)           ! observations channels
     integer,     intent(inout)                                :: KMARQ(KNO,KNT)           ! observations channels
     integer,     intent(in)                                   :: KNOSAT                   ! numero de satellite (i.e. indice) 
     integer,     intent(in)                                   :: KNO                      ! nombre de canaux des observations 
     integer,     intent(in)                                   :: KNT                      ! nombre de tovs
     character *9,intent(in)                                   :: STNID                    ! identificateur du satellite
-    integer,     intent(inout)                                :: B7CHCK(KNO,KNT)          ! 
+    integer,     intent(inout)                                :: B7CHCK(KNO,KNT) 
     integer,     intent(inout)                                :: ICHECK(KNO,KNT)          ! indicateur du QC par canal
     ! Locals
     integer                                                   :: nDataIndex
@@ -2911,7 +2911,7 @@ contains
             rejectionCodArray(testIndex,KCANO(nChannelIndex,nDataIndex),KNOSAT) = &
                  rejectionCodArray(testIndex,KCANO(nChannelIndex,nDataIndex),KNOSAT)+ 1
             if ( mwbg_debug ) then
-              write(*,*)STNID(2:9),' first bgckAtms program REJECT.', &
+              write(*,*)STNID(2:9),' first bgckAtms/bgckMwhs2 program REJECT.', &
                         'CHANNEL=', KCANO(nChannelIndex,nDataIndex), &
                         ' IMARQ= ',KMARQ(nChannelIndex,nDataIndex)
             end if
@@ -2920,19 +2920,18 @@ contains
       end do
     end if
 
-  end subroutine atmsTest1Flagbit7Check
+  end subroutine atmsMwhs2Test1Flagbit7Check
 
   !--------------------------------------------------------------------------
-  ! atmsTest2TopographyCheck
+  ! atmsMwhs2Test2TopographyCheck
   !--------------------------------------------------------------------------
-  subroutine atmsTest2TopographyCheck(itest, KCANO, KNOSAT, KNO, KNT, STNID, MTINTRP, &
-                                      KMARQ, ICHTOPO, MXTOPO, ZCRIT, B7CHCK, & 
-                                      ICHECK)
+  subroutine atmsMwhs2Test2TopographyCheck(itest, KCANO, KNOSAT, KNO, KNT, STNID, MTINTRP, &
+                                           KMARQ, ICHTOPO, MXTOPO, ZCRIT, B7CHCK, ICHECK)
 
     !:Purpose:               1) test 2: Topography check (partial)
 
     ! Arguments
-    integer,     intent(in)                                   :: itest(:)                 ! test number
+    integer,     intent(in)                                   :: itest(:)                       ! test number
     integer,     intent(in)                                   :: KCANO(KNO,KNT)                 ! observations channels
     integer,     intent(in)                                   :: KNOSAT                         ! numero de satellite (i.e. indice) 
     integer,     intent(in)                                   :: KNO                            ! nombre de canaux des observations 
@@ -2944,7 +2943,7 @@ contains
     integer,     intent(in)                                   :: ICHTOPO(MXTOPO) 
     real ,       intent(in)                                   :: ZCRIT(MXTOPO)
     integer,     intent(inout)                                :: ICHECK(KNO,KNT)                ! indicateur du QC par canal
-    integer,     intent(inout)                                :: B7CHCK(KNO,KNT)          ! 
+    integer,     intent(inout)                                :: B7CHCK(KNO,KNT)
     ! Locals
     integer                                                   :: nDataIndex
     integer                                                   :: nChannelIndex
@@ -2978,12 +2977,12 @@ contains
       end do
     end if
 
-  end subroutine atmsTest2TopographyCheck
+  end subroutine atmsMwhs2Test2TopographyCheck
 
   !--------------------------------------------------------------------------
-  ! atmsTest3UncorrectedTbCheck
+  ! atmsMwhs2Test3UncorrectedTbCheck
   !--------------------------------------------------------------------------
-  subroutine atmsTest3UncorrectedTbCheck(itest, KCANO, KNOSAT, KNO, KNT, STNID, RESETQC, &
+  subroutine atmsMwhs2Test3UncorrectedTbCheck(itest, KCANO, KNOSAT, KNO, KNT, STNID, RESETQC, &
                                          KMARQ, B7CHCK, ICHECK)
 
     !:Purpose:                       Test 3: Uncorrected Tb check (single)
@@ -2991,7 +2990,7 @@ contains
     !                                In this case switch bit 11 ON.
 
     ! Arguments
-    integer,     intent(in)                                   :: itest(:)                 ! test number
+    integer,     intent(in)                                   :: itest(:)                       ! test number
     integer,     intent(in)                                   :: KCANO(KNO,KNT)                 ! observations channels
     integer,     intent(in)                                   :: KNOSAT                         ! numero de satellite (i.e. indice) 
     integer,     intent(in)                                   :: KNO                            ! nombre de canaux des observations 
@@ -3000,7 +2999,7 @@ contains
     logical,     intent(in)                                   :: RESETQC                        ! resetqc logical
     integer,     intent(inout)                                :: KMARQ(KNO,KNT)                 ! marqueur de radiance 
     integer,     intent(inout)                                :: ICHECK(KNO,KNT)                ! indicateur du QC par canal
-    integer,     intent(inout)                                :: B7CHCK(KNO,KNT)          ! 
+    integer,     intent(inout)                                :: B7CHCK(KNO,KNT)
     ! Locals
     integer                                                   :: nDataIndex
     integer                                                   :: nChannelIndex
@@ -3035,7 +3034,7 @@ contains
       end if
     end if
 
-  end subroutine atmsTest3UncorrectedTbCheck
+  end subroutine atmsMwhs2Test3UncorrectedTbCheck
 
   !--------------------------------------------------------------------------
   ! atmsTest4RogueCheck
@@ -3073,12 +3072,12 @@ contains
     real,        intent(in)              :: clwFG(:)
     integer,     intent(in)              :: IDENTF(KNT)                    ! data flag ident  
     integer,     intent(in)              :: MXSFCREJ                       ! cst 
-    integer,     intent(in)              :: ISFCREJ(MXSFCREJ)              !
+    integer,     intent(in)              :: ISFCREJ(MXSFCREJ)
     integer,     intent(in)              :: MXCH2OMPREJ                    ! cst 
-    integer,     intent(in)              :: ICH2OMPREJ(MXCH2OMPREJ)        !
+    integer,     intent(in)              :: ICH2OMPREJ(MXCH2OMPREJ)
     integer,     intent(inout)           :: KMARQ(KNO,KNT)                 ! marqueur de radiance 
     integer,     intent(inout)           :: ICHECK(KNO,KNT)                ! indicateur du QC par canal
-    integer,     intent(inout)           :: B7CHCK(KNO,KNT)                ! 
+    integer,     intent(inout)           :: B7CHCK(KNO,KNT)
 
     ! Locals
     integer                              :: channelval
@@ -3203,12 +3202,150 @@ contains
       end do
     end if
 
-  end subroutine atmsTest4RogueCheck 
+  end subroutine atmsTest4RogueCheck
 
   !--------------------------------------------------------------------------
-  ! atmsTest5ChannelSelectionUsingIutilst
+  ! Mwhs2Test4RogueCheck
   !--------------------------------------------------------------------------
-  subroutine atmsTest5ChannelSelectionUsingIutilst(itest, KCANO, KNOSAT, KNO, KNT, STNID, &
+  subroutine Mwhs2Test4RogueCheck(itest, KCANO, KNOSAT, KNO, KNT, STNID, ROGUEFAC, TOVERRST, &
+                                  clwThreshArr, useStateDepSigmaObs, sigmaObsErr, waterobs, &
+                                  PTBOMP, clwObs, clwFG, IDENTF, ICH2OMPREJ, &
+                                  MXCH2OMPREJ, KMARQ, B7CHCK, ICHECK)
+
+    !:Purpose:                         test 4: "Rogue check" for (O-P) Tb residuals out of range.
+    !                                  (single/full).
+    !                                  Also, over WATER remove CH.10-15 if CH.10 |O-P|>5K (full)
+    !                                  Les observations, dont le residu (O-P)
+    !                                  depasse par un facteur (roguefac) l'erreur totale des TOVS.
+    !                                  OVER OPEN WATER
+    !                                  ch. 10 Abs(O-P) > 5K produces rejection of all ATMS amsub channels 10-15.
+
+    ! Arguments
+    integer,     intent(in)              :: itest(:)                       ! test number
+    integer,     intent(in)              :: KCANO(KNO,KNT)                 ! observations channels
+    integer,     intent(in)              :: KNOSAT                         ! numero de satellite (i.e. indice)
+    integer,     intent(in)              :: KNO                            ! nombre de canaux des observations
+    integer,     intent(in)              :: KNT                            ! nombre de tovs
+    character *9,intent(in)              :: STNID                          ! identificateur du satellite
+    real,        intent(in)              :: ROGUEFAC(:)                    ! rogue factor
+    real(8),     intent(in)              :: TOVERRST(:,:)                  ! erreur totale TOVs
+    logical,     intent(in)              :: useStateDepSigmaObs(:,:)       ! if using state dependent obs error
+    real(8),     intent(in)              :: clwThreshArr(:,:,:)            ! cloud threshold array
+    real(8),     intent(in)              :: sigmaObsErr(:,:,:)             ! sigma obs error
+    logical,     intent(in)              :: waterobs(:)                    ! open water obs
+    real,        intent(in)              :: PTBOMP(KNO,KNT)                ! radiance o-p
+    real,        intent(in)              :: clwObs(:)
+    real,        intent(in)              :: clwFG(:)
+    integer,     intent(in)              :: IDENTF(KNT)                    ! data flag ident
+    integer,     intent(in)              :: MXCH2OMPREJ                    ! cst
+    integer,     intent(in)              :: ICH2OMPREJ(MXCH2OMPREJ)
+    integer,     intent(inout)           :: KMARQ(KNO,KNT)                 ! marqueur de radiance
+    integer,     intent(inout)           :: ICHECK(KNO,KNT)                ! indicateur du QC par canal
+    integer,     intent(inout)           :: B7CHCK(KNO,KNT)
+
+    ! Locals
+    integer                              :: channelval
+    integer                              :: nDataIndex
+    integer                              :: nChannelIndex
+    integer                              :: testIndex
+    integer                              :: INDXCAN
+    real                                 :: XCHECKVAL
+    real                                 :: clwThresh1
+    real                                 :: clwThresh2
+    real                                 :: sigmaThresh1
+    real                                 :: sigmaThresh2
+    real                                 :: sigmaObsErrUsed
+    real                                 :: clwObsFGaveraged
+    logical                              :: CH2OMPREJCT
+    logical                              :: IBIT
+
+    testIndex = 4
+    if ( itest(testIndex) == 1 ) then
+      do nDataIndex=1,KNT
+        CH2OMPREJCT = .FALSE.
+        do nChannelIndex=1,KNO
+          channelval = KCANO(nChannelIndex,nDataIndex)
+          ! using state-dependent obs error only over water.
+          ! obs over sea-ice will be rejected in test 15.
+          if ( tvs_mwAllskyAssim .and. useStateDepSigmaObs(channelval,KNOSAT) &
+                .and. waterobs(nDataIndex) ) then
+            clwThresh1 = clwThreshArr(channelval,KNOSAT,1)
+            clwThresh2 = clwThreshArr(channelval,KNOSAT,2)
+            sigmaThresh1 = sigmaObsErr(channelval,KNOSAT,1)
+            sigmaThresh2 = sigmaObsErr(channelval,KNOSAT,2)
+            clwObsFGaveraged = 0.5 * (clwObs(nDataIndex) + clwFG(nDataIndex))
+            if ( clwObs(nDataIndex) == mwbg_realMissing ) then
+              sigmaObsErrUsed = MPC_missingValue_R4
+            else
+              sigmaObsErrUsed = calcStateDepObsErr_r4(clwThresh1,clwThresh2,sigmaThresh1, &
+                                                        sigmaThresh2,clwObsFGaveraged)
+            end if
+          else
+            sigmaObsErrUsed = TOVERRST(channelval,KNOSAT)
+          end if
+          ! For sigmaObsErrUsed=MPC_missingValue_R4 (clwObs=mwbg_realMissing
+          ! in all-sky mode), the observation is flagged for rejection in
+          ! mwbg_reviewAllCritforFinalFlagsMwhs2.
+          XCHECKVAL = ROGUEFAC(channelval) * sigmaObsErrUsed
+          if ( PTBOMP(nChannelIndex,nDataIndex)      /= mwbg_realMissing    .and. &
+              ABS(PTBOMP(nChannelIndex,nDataIndex))  >= XCHECKVAL .and. &
+              sigmaObsErrUsed /= MPC_missingValue_R4 ) then
+            ICHECK(nChannelIndex,nDataIndex) = MAX(ICHECK(nChannelIndex,nDataIndex),testIndex)
+            KMARQ(nChannelIndex,nDataIndex) = OR(KMARQ(nChannelIndex,nDataIndex),2**9)
+            KMARQ(nChannelIndex,nDataIndex) = OR(KMARQ(nChannelIndex,nDataIndex),2**16)
+            rejectionCodArray(testIndex,KCANO(nChannelIndex,nDataIndex),KNOSAT) =  &
+               rejectionCodArray(testIndex,KCANO(nChannelIndex,nDataIndex),KNOSAT) + 1
+            if ( B7CHCK(nChannelIndex,nDataIndex) == 0 ) then
+              rejectionCodArray2(testIndex,KCANO(nChannelIndex,nDataIndex),KNOSAT) = &
+                 rejectionCodArray2(testIndex,KCANO(nChannelIndex,nDataIndex),KNOSAT)+ 1
+            end if
+            if ( mwbg_debug ) then
+              write(*,*)STNID(2:9),'ROGUE CHECK REJECT.NO.', &
+                     ' OBS = ',nDataIndex, &
+                     ' CHANNEL= ',KCANO(nChannelIndex,nDataIndex), &
+                     ' CHECK VALUE= ',XCHECKVAL, &
+                     ' TBOMP= ',PTBOMP(nChannelIndex,nDataIndex), &
+                     ' TOVERRST= ',TOVERRST(channelval,KNOSAT)
+            end if
+          end if
+          if ( channelval == 10 .and. PTBOMP(nChannelIndex,nDataIndex) /= mwbg_realMissing .and. &
+              ABS(PTBOMP(nChannelIndex,nDataIndex)) > 5.0 ) then
+            CH2OMPREJCT = .TRUE.
+          end if
+        end do
+
+        !    Channels 10-15 are rejected if, for ch10 ABS(O-P) > 5K
+        !    Apply over open water only (bit 0 ON in QC integer identf).
+        !    Only apply if obs not rejected in this test already.
+        IBIT = AND(IDENTF(nDataIndex), 2**0)
+        if ( CH2OMPREJCT .and. (IBIT /= 0) ) then
+          do nChannelIndex=1,KNO
+            INDXCAN = ISRCHEQI (ICH2OMPREJ,MXCH2OMPREJ,KCANO(nChannelIndex,nDataIndex))
+            if ( INDXCAN /= 0 )  then
+              if ( ICHECK(nChannelIndex,nDataIndex) /= testIndex ) then
+                ICHECK(nChannelIndex,nDataIndex) = MAX(ICHECK(nChannelIndex,nDataIndex),testIndex)
+                KMARQ(nChannelIndex,nDataIndex) = OR(KMARQ(nChannelIndex,nDataIndex),2**9)
+                KMARQ(nChannelIndex,nDataIndex) = OR(KMARQ(nChannelIndex,nDataIndex),2**16)
+                rejectionCodArray(testIndex,KCANO(nChannelIndex,nDataIndex),KNOSAT) = &
+                        rejectionCodArray(testIndex,KCANO(nChannelIndex,nDataIndex),KNOSAT)+ 1
+                if ( B7CHCK(nChannelIndex,nDataIndex) == 0 ) then
+                  rejectionCodArray2(testIndex,KCANO(nChannelIndex,nDataIndex),KNOSAT) = &
+                     rejectionCodArray2(testIndex,KCANO(nChannelIndex,nDataIndex),KNOSAT)+ 1
+                end if
+              end if
+            end if
+          end do
+        end if
+
+      end do
+    end if
+
+  end subroutine Mwhs2Test4RogueCheck
+
+  !--------------------------------------------------------------------------
+  ! atmsMwhs2Test5ChannelSelectionUsingIutilst
+  !--------------------------------------------------------------------------
+  subroutine atmsMwhs2Test5ChannelSelectionUsingIutilst(itest, KCANO, KNOSAT, KNO, KNT, STNID, &
                                                    IUTILST, KMARQ, ICHECK)
 
     !:Purpose:                         test 5: Channel selection using array IUTILST(chan,sat)
@@ -3255,7 +3392,7 @@ contains
       write(*,*) 'ICHECK = ',((ICHECK(nChannelIndex,nDataIndex),nChannelIndex=1,KNO),nDataIndex=1,KNT)
     end if
 
-  end subroutine atmsTest5ChannelSelectionUsingIutilst
+  end subroutine atmsMwhs2Test5ChannelSelectionUsingIutilst
 
   !--------------------------------------------------------------------------
   ! mwbg_tovCheckAtms 
@@ -3422,7 +3559,7 @@ contains
     ! STEP 2 ) Check for values of TB that are missing or outside physical limits.
     !###############################################################################
 
-    call mwbg_grossValueCheck(KNT, KNO, ztb, 50., 380., grossrej)
+    call mwbg_grossValueCheck(KNT, KNO, ztb, biasCorr, 50., 380., grossrej)
 
     !###############################################################################
     ! STEP 3 ) Preliminary QC checks --> set lqc(KNT,KNO)=.true.
@@ -3499,16 +3636,16 @@ contains
     ! 1) test 1: Check flag bit 7 on from the first bgckAtms program
     !  Includes observations flagged for cloud liquid water, scattering index,
     !  dryness index plus failure of several QC checks.
-    call atmsTest1Flagbit7Check (itest, KCANO, KMARQ, KNOSAT, ICHECK, KNO, KNT, &
-                                 STNID,  B7CHCK)
+    call atmsMwhs2Test1Flagbit7Check (itest, KCANO, KMARQ, KNOSAT, ICHECK, KNO, KNT, &
+                                      STNID,  B7CHCK)
 
     ! 2) test 2: Topography check (partial)
-    call atmsTest2TopographyCheck (itest, KCANO, KNOSAT, KNO, KNT, STNID, MTINTRP, &
-                                   KMARQ, ICHTOPO, MXTOPO, ZCRIT, B7CHCK, ICHECK)
+    call atmsMwhs2Test2TopographyCheck (itest, KCANO, KNOSAT, KNO, KNT, STNID, MTINTRP, &
+                                        KMARQ, ICHTOPO, MXTOPO, ZCRIT, B7CHCK, ICHECK)
     ! 3) test 3: Uncorrected Tb check (single)
     !  Uncorrected datum (flag bit #6 off). In this case switch bit 11 ON.
-    call atmsTest3UncorrectedTbCheck (itest, KCANO, KNOSAT, KNO, KNT, STNID, RESETQC, &
-                                      KMARQ, B7CHCK, ICHECK)
+    call atmsMwhs2Test3UncorrectedTbCheck (itest, KCANO, KNOSAT, KNO, KNT, STNID, RESETQC, &
+                                           KMARQ, B7CHCK, ICHECK)
     ! 4) test 4: "Rogue check" for (O-P) Tb residuals out of range. (single/full)
     !             Also, over WATER remove CH.17-22 if CH.17 |O-P|>5K (partial)
     !  Les observations, dont le residu (O-P) depasse par un facteur (roguefac)
@@ -3525,8 +3662,8 @@ contains
     ! 5) test 5: Channel selection using array IUTILST(chan,sat)
     !  IUTILST = 0 (blacklisted)
     !            1 (assmilate)
-    call atmsTest5ChannelSelectionUsingIutilst(itest, KCANO, KNOSAT, KNO, KNT, STNID, &
-                                               IUTILST, KMARQ, ICHECK)
+    call atmsMwhs2Test5ChannelSelectionUsingIutilst(itest, KCANO, KNOSAT, KNO, KNT, STNID, &
+                                                    IUTILST, KMARQ, ICHECK)
 
     !  Synthese de la controle de qualite au niveau de chaque point
     !  d'observation. Code:
@@ -3599,17 +3736,27 @@ contains
   !--------------------------------------------------------------------------
   ! mwbg_tovCheckMwhs2
   !--------------------------------------------------------------------------
-  subroutine mwbg_tovCheckMwhs2(zlat, zlon, ilq, itt, zenith, &
+  subroutine mwbg_tovCheckMwhs2(TOVERRST, clwThreshArr, sigmaObsErr, useStateDepSigmaObs, &
+                                IUTILST, zlat, zlon, ilq, itt, zenith, &
                                 ICANO, ztb, biasCorr, ZOMP, ICHECK, KNO, KNT, KNOSAT, IDENT, &
-                                ISCNPOS, globMarq, IMARQ, clwObs, clwFG, riwv, &
-                                RESETQC, modLSQ)
+                                ISCNPOS, MTINTRP, globMarq, IMARQ, clwObs, clwFG, riwv, &
+                                STNID, RESETQC, modLSQ, lastHeader)
 
 
     !:Purpose:                   Effectuer le controle de qualite des radiances tovs.
     !
 
     implicit none
+    !Arguments
+    integer, intent(in)              :: IUTILST(:,:)           ! channel Selection using array IUTILST(chan,sat)
+    !                                                            IUTILST = 0 (blacklisted)
+    !                                                            1 (assmilate)
+    !                                                            2 (assimilate over open water only)
 
+    real(8), intent(in)              :: TOVERRST(:,:)          ! l'erreur totale des TOVS
+    real(8), intent(in)              :: clwThreshArr(:,:,:)
+    real(8), intent(in)              :: sigmaObsErr(:,:,:)
+    logical, intent(in)              :: useStateDepSigmaObs(:,:) ! if using state dependent obs error
     integer, intent(in)              :: KNO                    ! nombre de canaux des observations
     integer, intent(in)              :: KNT                    ! nombre de tovs
     real,    intent(in)              :: zlat(:)
@@ -3624,11 +3771,14 @@ contains
     real, intent(inout)              :: ztb(:)                 ! radiances
     real, intent(in)                 :: biasCorr(:)            ! correction aux radiances
     real, intent(in)                 :: zomp(:)                ! residus (o-p)
+    real, intent(in)                 :: MTINTRP(KNT)           ! topographie du modele
     integer, allocatable, intent(out):: IDENT(:)               ! flag to identify all obs pts in report
     !                                                            as being over land/ice, cloudy, bad IWV
+    character *9, intent(in)         :: STNID                  ! identificateur du satellite
     logical, intent(in)              :: RESETQC                ! reset du controle de qualite?
     logical, intent(in)              :: modLSQ                 ! If active, recalculate values for land/sea
                                                                ! qualifier and terrain type based on LG/MG
+    logical, intent(in)              :: lastHeader             ! active if last header
     integer,allocatable, intent(out) :: ICHECK(:,:)            ! indicateur controle de qualite tovs par canal
     !                                                            =0, ok,
     !                                                            >0, rejet,
@@ -3672,12 +3822,15 @@ contains
     integer                          :: kk
     integer                          :: INDX
     integer                          :: ICH2OMPREJ(MXCH2OMPREJ)
+    integer                          :: B7CHCK(KNO,KNT)
     real, allocatable                :: ROGUEFAC(:)
     real                             :: ZCRIT(MXTOPO)
+    integer                          :: ITEST(mwbg_maxNumTest)
     integer                          :: chanFlaggedForAllskyGenCoeff(MXCLWREJ)
     integer                          :: ICHTOPO(MXTOPO)
     logical, save                    :: LLFIRST = .true.
     integer, save                    :: numReportWithMissingTb
+    integer, save                    :: allcnt                          ! Number of Tovs obs
     integer, save                    :: drycnt                          ! Number of pts flagged for AMSU-B Dryness Index
     integer, save                    :: landcnt                         ! Number of obs pts found over land/ice
     integer, save                    :: rejcnt                          ! Number of problem obs pts (Tb err, QCfail)
@@ -3705,12 +3858,18 @@ contains
     ICHTOPO(:) = (/11, 12, 13/)
     ZCRIT(:) = (/2500., 2250., 2000./)
 
+    !  Test selection (0=skip test, 1=do test)
+    !              1  2  3  4  5
+    ITEST(:)  = 0
+    ITEST(1:5) = (/1, 1, 1, 1, 1/)
+
     ! Channels excluded from gen_bias_corr in all-sky mode
     chanFlaggedForAllskyGenCoeff(:) = (/ 10, 11, 12, 13, 14, 15/)
 
     ! Initialisation, la premiere fois seulement!
     if (LLFIRST) then
       numReportWithMissingTb = 0
+      allcnt = 0
       flgcnt = 0
       landcnt = 0
       rejcnt = 0
@@ -3736,17 +3895,14 @@ contains
     !###############################################################################
     ! STEP 2 ) Check for values of TB that are missing or outside physical limits.
     !###############################################################################
-
-    call mwbg_grossValueCheck(KNT, KNO, ztb, 50., 380., grossrej)
+    call mwbg_grossValueCheck(KNT, KNO, ztb, biasCorr, 50., 380., grossrej)
 
     !###############################################################################
     ! STEP 3 ) Preliminary QC checks --> set lqc(KNT,KNO)=.true.
     !          for data that fail QC
     !###############################################################################
-
     call mwbg_firstQcCheckMwhs2(zenith, ilq, itt, zlat, zlon, ztb, ISCNPOS, &
-                                KNO, KNT, lqc, lsq, trn, &
-                                ICANO, reportHasMissingTb, modLSQ)
+                                KNO, KNT, lqc, lsq, trn, ICANO, reportHasMissingTb, modLSQ)
 
     if ( reportHasMissingTb ) numReportWithMissingTb = numReportWithMissingTb + 1
     !  Exclude problem points from further calculations
@@ -3791,7 +3947,7 @@ contains
     call mwbg_reviewAllCritforFinalFlagsMwhs2(KNT, KNO, lqc, grossrej, trn, waterobs, &
                                               precipobs, clwObs, clwFG, scatec, scatbg, &
                                               iwvreject, riwv, IMARQ, globMarq, zdi, ident, &
-                                              drycnt, landcnt, rejcnt, iwvcnt, pcpcnt, flgcnt, &
+                                              allcnt, drycnt, landcnt, rejcnt, iwvcnt, pcpcnt, flgcnt, &
                                               MXCLWREJ, chanFlaggedForAllskyGenCoeff, icano)
 
     !###############################################################################
@@ -3807,14 +3963,50 @@ contains
     call utl_reAllocate(icheck, KNO, KNT)
     !  Initialisations
     ICHECK(:,:) = 0
+    B7CHCK(:,:) = 0
 
     if ( RESETQC ) KMARQ(:,:) = 0
+
+    ! 1) test 1: Check flag bit 7 on from the first bgckMwhs2 program
+    !  Includes observations flagged for cloud liquid water, scattering index,
+    !  dryness index plus failure of several QC checks.
+    call atmsMwhs2Test1Flagbit7Check (itest, KCANO, KMARQ, KNOSAT, ICHECK, KNO, KNT, &
+                                      STNID,  B7CHCK)
+
+    ! 2) test 2: Topography check (partial)
+    call atmsMwhs2Test2TopographyCheck (itest, KCANO, KNOSAT, KNO, KNT, STNID, MTINTRP, &
+                                        KMARQ, ICHTOPO, MXTOPO, ZCRIT, B7CHCK, ICHECK)
+    ! 3) test 3: Uncorrected Tb check (single)
+    !  Uncorrected datum (flag bit #6 off). In this case switch bit 11 ON.
+    call atmsMwhs2Test3UncorrectedTbCheck (itest, KCANO, KNOSAT, KNO, KNT, STNID, RESETQC, &
+                                           KMARQ, B7CHCK, ICHECK)
+    ! 4) test 4: "Rogue check" for (O-P) Tb residuals out of range. (single/full)
+    !             Also, over WATER remove CH.10-15 if CH.10 |O-P|>5K (full)
+    !  Les observations, dont le residu (O-P) depasse par un facteur (roguefac)
+    !   l'erreur totale des TOVS.
+    !  OVER OPEN WATER
+    !    ch. 10 Abs(O-P) > 5K produces rejection of all ATMS amsub channels 10-15.
+    call Mwhs2Test4RogueCheck (itest, KCANO, KNOSAT, KNO, KNT, STNID, ROGUEFAC, TOVERRST, &
+                              clwThreshArr, useStateDepSigmaObs, sigmaObsErr, waterobs, &
+                              PTBOMP, clwObs, clwFG, IDENT, ICH2OMPREJ, &
+                              MXCH2OMPREJ, KMARQ, B7CHCK, ICHECK)
+
+    ! 5) test 5: Channel selection using array IUTILST(chan,sat)
+    !  IUTILST = 0 (blacklisted)
+    !            1 (assmilate)
+    call atmsMwhs2Test5ChannelSelectionUsingIutilst(itest, KCANO, KNOSAT, KNO, KNT, STNID, &
+                                                    IUTILST, KMARQ, ICHECK)
 
     !  Synthese de la controle de qualite au niveau de chaque point
     !  d'observation. Code:
     !            =0, aucun rejet,
     !            >0, au moins un canal rejete.
-    KCHKPRF(:) = 0
+    do JJ=1,KNT
+      KCHKPRF(JJ) = 0
+      do JI=1,KNO
+        KCHKPRF(JJ) = MAX(KCHKPRF(JJ),ICHECK(JI,JJ))
+      end do
+    end do
 
     if ( mwbg_debug ) then
       write(*,*)'KCHKPRF = ',(KCHKPRF(JJ),JJ=1,KNT)
@@ -3831,8 +4023,9 @@ contains
     ! reset global marker flag (55200) and mark it if observtions are rejected
     call resetQcCases(RESETQC, KCHKPRF, globMarq)
 
-    if(mwbg_debug) then
+    if (lastHeader) then
       write(*,*) ' --------------------------------------------------------------- '
+      write(*,*) ' Number of obs pts read from BURP file                         = ', allcnt
       write(*,*) ' Number of BURP file reports where Tb set to mwbg_realMissing  = ', numReportWithMissingTb
       write(*,*) ' --------------------------------------------------------------- '
       write(*,*) ' 1. Number of obs pts found over land/ice           = ', landcnt
@@ -4773,7 +4966,7 @@ contains
   ! mwbg_grossValueCheck  
   !--------------------------------------------------------------------------
 
-  subroutine mwbg_grossValueCheck(npts,KNO, ztb, ztbThresholdMin, ztbThresholdMax, grossrej)
+  subroutine mwbg_grossValueCheck(npts, KNO, ztbcor, biasCorr, ztbThresholdMin, ztbThresholdMax, grossrej)
 
     !:Purpose: Check Tbs for values that are missing or outside physical limits.
     !          **NOTE: REJECT ALL CHANNELS OF ONE IS FOUND TO BE BAD.
@@ -4782,25 +4975,40 @@ contains
     ! Arguments
     integer, intent(in)               :: npts             ! number of obs pts to process
     integer, intent(in)               :: KNO              ! number of ichannels
-    real,    intent(in)               :: ztb(:)           ! bs from input BURP file
+    real,    intent(in)               :: ztbcor(:)        ! bs from input BURP file
+    real,    intent(in)               :: biasCorr(:)      ! Bias correction
     real,    intent(in)               :: ztbThresholdMin  ! ztb threshold for rejection
     real,    intent(in)               :: ztbThresholdMax  ! ztb threshold for rejection
-    logical, intent(out), allocatable :: grossrej(:)      ! ogical array defining which obs are to be rejected
+    logical, intent(out), allocatable :: grossrej(:)      ! logical array defining which obs are to be rejected
 
     ! Locals
-    integer :: ii, indx1, indx2
+    integer :: pt, indx, channel
 
+    real, allocatable                 :: ztb(:)           ! biased or unbiased radiances
+
+    call utl_reAllocate(ztb, KNO)
     call utl_reAllocate(grossrej, npts)
     
     grossrej(1:npts) = .true.
-    indx1 = 1
-    do ii = 1, npts
-
-      indx2 = ii*KNO
-      if ( all( ztb(indx1:indx2) > ztbThresholdMin ) .and. all( ztb(indx1:indx2) < ztbThresholdMax ) ) then
-        grossrej(ii) = .false.
+    indx = 0
+    do pt = 1, npts
+      if ( mwbg_useUnbiasedObsForClw ) then
+        do channel = 1, KNO
+          ztb(channel) = ztbcor(indx+channel)
+        end do
+      else
+        do channel = 1, KNO
+          if (biasCorr(indx+channel) /= mwbg_realMissing) then
+            ztb(channel) = ztbcor(indx+channel) - biasCorr(indx+channel)
+          else
+            ztb(channel) = ztbcor(indx+channel)
+          end if
+        end do
       end if
-      indx1 = indx2 + 1
+      if ( all( ztb > ztbThresholdMin ) .and. all( ztb < ztbThresholdMax ) ) then
+        grossrej(pt) = .false.
+      end if
+      indx = pt*KNO
 
     end do
 
@@ -5047,8 +5255,7 @@ contains
   ! mwbg_firstQcCheckMwhs2
   !--------------------------------------------------------------------------
   subroutine mwbg_firstQcCheckMwhs2(zenith, ilq, itt, zlat, zlon, ztb, scanpos, &
-                                    nval, nt, lqc, lsq, trn, &
-                                    ican, reportHasMissingTb, modLSQ)
+                                    nval, nt, lqc, lsq, trn, ican, reportHasMissingTb, modLSQ)
     !  This routine performs basic quality control checks on the data. It sets array
     !  lqc(nt,nval) elements to .true. to flag data with failed checks. Check 1
     !  (for ilq,itt) and check 5 are skipped if modlsqtt=.true., as the original values
@@ -5367,6 +5574,9 @@ contains
     call utl_reAllocate(si_ecmwf,ni)
     call utl_reAllocate(si_bg,ni)
     call utl_reAllocate(SeaIce,ni)
+
+    iNumSeaIce = 0
+
     ! extract required channels:
     !  23 Ghz = AMSU-A 1 = ATMS channel 1
     !  31 Ghz = AMSU-A 2 = ATMS channel 2
@@ -5431,19 +5641,32 @@ contains
         abslat = abs(plat(i))
         cosz   = cosd(pangl(i))
 
-        if ( mwbg_useUnbiasedObsForClw ) then
+        if (bcor23(i) == mwbg_realMissing .or. mwbg_useUnbiasedObsForClw) then
           t23 = tb23(i)
-          t31 = tb31(i)
-          t50 = tb50(i)
-          t89 = tb89(i)
-          t165 = tb165(i)
         else
           t23 = tb23(i) - bcor23(i)
+        end if
+        if (bcor31(i) == mwbg_realMissing .or. mwbg_useUnbiasedObsForClw) then
+          t31 = tb31(i)
+        else
           t31 = tb31(i) - bcor31(i)
+        end if
+        if (bcor50(i) == mwbg_realMissing .or. mwbg_useUnbiasedObsForClw) then
+          t50 = tb50(i)
+        else
           t50 = tb50(i) - bcor50(i)
+        end if
+        if (bcor89(i) == mwbg_realMissing .or. mwbg_useUnbiasedObsForClw) then
+          t89 = tb89(i)
+        else
           t89 = tb89(i) - bcor89(i)
+        end if
+        if (bcor165(i) == mwbg_realMissing .or. mwbg_useUnbiasedObsForClw) then
+          t165 = tb165(i)
+        else
           t165 = tb165(i) - bcor165(i)
         end if
+
         deltb = t89 - t165
         t23FG = tb23FG(i)
         t31FG = tb31FG(i)
@@ -5620,6 +5843,8 @@ contains
     call utl_reAllocate(si_bg,ni)
     call utl_reAllocate(SeaIce,ni)
 
+    iNumSeaIce = 0
+
     !! extract required channels:  ATMS ch.   MWHS-2 ch.
     !!   23 Ghz = AMSU-A 1 =        1          n/a
     !!   31 Ghz = AMSU-A 2 =        2          n/a
@@ -5688,19 +5913,32 @@ contains
         abslat = abs(plat(i))
         cosz   = cosd(pangl(i))
 
-        if ( mwbg_useUnbiasedObsForClw ) then
+        if (bcor23(i) == mwbg_realMissing .or. mwbg_useUnbiasedObsForClw) then
           t23 = tb23(i)
-          t31 = tb31(i)
-          t50 = tb50(i)
-          t89 = tb89(i)
-          t165 = tb165(i)
         else
           t23 = tb23(i) - bcor23(i)
+        end if
+        if (bcor31(i) == mwbg_realMissing .or. mwbg_useUnbiasedObsForClw) then
+          t31 = tb31(i)
+        else
           t31 = tb31(i) - bcor31(i)
+        end if
+        if (bcor50(i) == mwbg_realMissing .or. mwbg_useUnbiasedObsForClw) then
+          t50 = tb50(i)
+        else
           t50 = tb50(i) - bcor50(i)
+        end if
+        if (bcor89(i) == mwbg_realMissing .or. mwbg_useUnbiasedObsForClw) then
+          t89 = tb89(i)
+        else
           t89 = tb89(i) - bcor89(i)
+        end if
+        if (bcor165(i) == mwbg_realMissing .or. mwbg_useUnbiasedObsForClw) then
+          t165 = tb165(i)
+        else
           t165 = tb165(i) - bcor165(i)
         end if
+
         deltb = t89 - t165
         t23FG = tb23FG(i)
         t31FG = tb31FG(i)
@@ -5715,7 +5953,6 @@ contains
           end if
 
           SeaIce(i) = ice(i)
-
           if ( ice(i) >= 0.55 .and. waterobs(i) ) then
             iNumSeaIce = iNumSeaIce + 1
             waterobs(i) = .false.
@@ -5804,6 +6041,7 @@ contains
     real, allocatable, intent(out)             :: riwv(:)
 
     ! Locals
+    integer                                    :: indx
     integer                                    :: indx1
     integer                                    :: indx2
     integer                                    :: ii
@@ -5853,19 +6091,13 @@ contains
     do ii = 1, nt
       indx2 = ii*nval
       if (.not.grossrej(ii)) then
-        if ( useUnbiasedObsForClw ) then
-          ztb183(1) = ztbcor(indx1+17)
-          ztb183(2) = ztbcor(indx1+18)
-          ztb183(3) = ztbcor(indx1+19)
-          ztb183(4) = ztbcor(indx1+20)
-          ztb183(5) = ztbcor(indx1+21)
-        else
-          ztb183(1) = ztbcor(indx1+17) - biasCorr(indx1+17)
-          ztb183(2) = ztbcor(indx1+18) - biasCorr(indx1+18)
-          ztb183(3) = ztbcor(indx1+19) - biasCorr(indx1+19)
-          ztb183(4) = ztbcor(indx1+20) - biasCorr(indx1+20)
-          ztb183(5) = ztbcor(indx1+21) - biasCorr(indx1+21)
-        end if
+        do indx = 1, 5
+          if (biasCorr(indx1+indx+9) == mwbg_realMissing .or. useUnbiasedObsForClw) then
+            ztb183(indx) = ztbcor(indx1+indx+16)
+          else
+            ztb183(indx) = ztbcor(indx1+indx+16) - biasCorr(indx1+indx+16)
+          end if
+        end do
         riwv(ii)  = sum(ztb183)/5.0
         if ( riwv(ii) < mean_Tb_183Ghz_min ) iwvreject(ii) = .true.
       else
@@ -5926,8 +6158,6 @@ contains
     !                                - 4     scatec > Lower Troposphere limit 9/10 (precipobs)
     !                                - 5     Mean 183 Ghz [ch. 18-22] Tb < 240K
     !                                - 6     CLW > clw_mwhs2_nrl_UTrej (0.200 kg/m2)
-    !                                - 7     Dryness Index rejection (for ch. 22)
-    !                                - 9     Dryness Index rejection (for ch. 21)
     !                               - 10     Sea ice > 0.55 detected
     !                               - 11     Gross error in Tb (any chan.)  (all channels rejected)
 
@@ -5952,6 +6182,7 @@ contains
     real, allocatable, intent(out)             :: riwv(:)
 
     ! Locals
+    integer                                    :: indx
     integer                                    :: indx1
     integer                                    :: indx2
     integer                                    :: ii
@@ -6001,19 +6232,13 @@ contains
     do ii = 1, nt
       indx2 = ii*nval
       if (.not.grossrej(ii)) then
-        if ( useUnbiasedObsForClw ) then
-          ztb183(1) = ztbcor(indx1+10)
-          ztb183(2) = ztbcor(indx1+11)
-          ztb183(3) = ztbcor(indx1+12)
-          ztb183(4) = ztbcor(indx1+13)
-          ztb183(5) = ztbcor(indx1+14)
-        else
-          ztb183(1) = ztbcor(indx1+10) - biasCorr(indx1+10)
-          ztb183(2) = ztbcor(indx1+11) - biasCorr(indx1+11)
-          ztb183(3) = ztbcor(indx1+12) - biasCorr(indx1+12)
-          ztb183(4) = ztbcor(indx1+13) - biasCorr(indx1+13)
-          ztb183(5) = ztbcor(indx1+14) - biasCorr(indx1+14)
-        end if
+        do indx = 1, 5
+          if (biasCorr(indx1+indx+9) == mwbg_realMissing .or. useUnbiasedObsForClw) then
+            ztb183(indx) = ztbcor(indx1+indx+9)
+          else
+            ztb183(indx) = ztbcor(indx1+indx+9) - biasCorr(indx1+indx+9)
+          end if
+        end do
         riwv(ii)  = sum(ztb183)/5.0
         if ( riwv(ii) < mean_Tb_183Ghz_min ) iwvreject(ii) = .true.
       else
@@ -6043,13 +6268,13 @@ contains
 
     ! Compute the simple AMSU-B Dryness Index zdi for all points = Tb(ch.3)-Tb(ch.5)
     if ( useUnbiasedObsForClw ) then
-      where ( .not.grossrej )
+      where ( .not. grossrej )
         zdi = ztb_amsub3 - ztb_amsub5
       elsewhere
         zdi = mwbg_realMissing
       end where
     else
-      where ( .not.grossrej )
+      where ( .not. grossrej )
         zdi = (ztb_amsub3 - bcor_amsub3) - (ztb_amsub5 - bcor_amsub5)
       elsewhere
         zdi = mwbg_realMissing
@@ -6251,7 +6476,7 @@ contains
   subroutine mwbg_reviewAllCritforFinalFlagsMwhs2(nt, nval, lqc, grossrej, trn, waterobs, &
                                                   precipobs, clwObs, clwFG, scatec, scatbg, &
                                                   iwvreject, riwv, IMARQ, globMarq, zdi, ident, &
-                                                  drycnt, landcnt, rejcnt, iwvcnt, pcpcnt, flgcnt, &
+                                                  allcnt, drycnt, landcnt, rejcnt, iwvcnt, pcpcnt, flgcnt, &
                                                   MXCLWREJ, chanFlaggedForAllskyGenCoeff, icano)
 
     !:Purpose:                   Review all the checks previously made to determine which obs are to be accepted
@@ -6278,6 +6503,7 @@ contains
     real, intent(inout)                        :: riwv(:)
     integer, intent(inout)                     :: IMARQ(:)
     integer, intent(inout)                     :: globMarq(:)
+    integer, intent(inout)                     :: allcnt
     integer, intent(inout)                     :: drycnt
     integer, intent(inout)                     :: landcnt
     integer, intent(inout)                     :: rejcnt
@@ -6301,6 +6527,7 @@ contains
 
     lflagchn(:,:) = lqc(:,:)  ! initialize with flags set in mwbg_firstQcCheckMwhs2
     do kk = 1, nt
+      allcnt = allcnt + 1  ! Counting total number of observations
       ! Reject all channels if gross Tb error detected in any channel or other problems
       if ( grossrej(kk) ) then
         lflagchn(kk,:) = .true.
@@ -6719,6 +6946,7 @@ contains
     real,    allocatable          :: atmScatteringIndex(:)         ! scattering index
     integer, external             :: exdb, exfin, fnom, fclos
     logical                       :: mwDataPresent
+    logical                       :: lastHeader                    ! active while reading last report
 
 
     call utl_tmg_start(118,'--BgckMicrowave')
@@ -6750,11 +6978,13 @@ contains
     !
     ! loop over all header indices of the specified family with surface obs
     numObsToProcess = 1
+    lastHeader = .false.
 
     call obs_set_current_header_list(obsSpaceData,'TO')
     HEADER: do
       headerIndex = obs_getHeaderIndex(obsSpaceData)
       if (headerIndex < 0) exit HEADER
+      if (headerIndex == obs_numHeader(obsSpaceData)) lastHeader = .true.
       codtyp = obs_headElem_i(obsSpaceData, OBS_ITY, headerIndex)
       if (instName== 'AMSUB') then
         if ( .not. tvs_isIdBurpInst(codtyp,'amsub') .and. &
@@ -6822,14 +7052,15 @@ contains
                                cloudLiquidWaterPathObs, cloudLiquidWaterPathFG, &
                                atmScatteringIndex, burpFileSatId, RESETQC)
       else if (instName == 'MWHS2') then
-        call mwbg_tovCheckMwhs2(obsLatitude, obsLongitude,&
-                                landQualifierIndice, terrainTypeIndice, satZenithAngle,   &
+        call mwbg_tovCheckMwhs2(oer_toverrst, oer_clwThreshArr, oer_sigmaObsErr, oer_useStateDepSigmaObs, &
+                                oer_tovutil, obsLatitude, obsLongitude, &
+                                landQualifierIndice, terrainTypeIndice, satZenithAngle,  &
                                 obsChannels, obsTb, obsTbBiasCorr, ompTb, qcIndicator,   &
                                 actualNumChannel, numObsToProcess, sensorIndex,          &
                                 newInformationFlag, satScanPosition,   &
-                                obsGlobalMarker, obsFlags,            &
+                                modelInterpTerrain, obsGlobalMarker, obsFlags,            &
                                 cloudLiquidWaterPathObs, cloudLiquidWaterPathFG, &
-                                atmScatteringIndex, RESETQC, modLSQ)
+                                atmScatteringIndex, burpFileSatId, RESETQC, modLSQ, lastHeader)
       else
         write(*,*) 'midas-bgckMW: instName = ', instName
         call utl_abort('midas-bgckMW: unknown instName')